+++ docs/components/grid/Grid.svelte
@@ -40,19 +40,20 @@ function onclick (e) {
 </script>
 `}</Code>
 <API props={apiProps}/>
+<API title="GridColumn Props" props={columnProps}/>
 
 
 <script lang="ts">
 import './Grid.css';
 import type { ApiProp } from '../../api-table/types';
 import { API, PROPS } from '../../api-table';
-import { Grid } from '../../../src/grid';
+import { Grid, type GridColumn } from '../../../src/grid';
 import { Code } from '../../code-example';
 
 const apiProps = <ApiProp[]>[
 	...PROPS.component,
-	{ name: 'columns', type: 'array', description: 'Array of columns definition. Each objects defines and describes a column in the grid.' },
-	{ name: 'data', type: 'array', description: 'Array of items (each would constitute 1 row of the grid).' },
+	{ name: 'columns', type: 'GridColumn[]', description: 'Array of columns definition. Each objects defines and describes a column in the grid.' },
+	{ name: 'data', type: 'any[]', description: 'Array of items (each would constitute 1 row of the grid).' },
 	{ name: 'round', description: 'Adds rounded corners to the table.' },
 	{ name: 'rowSelector', type: 'string', default: '<em>tbody tr</em>', description: 'A selector for a table row.<br>This is useful if a table needs row groups, in which case it would have a mix of TRs and TBODYs for rows. Both can have the same class, e.g. <em>.row</em> and this selector should then be provided here.' },
 	{ name: 'scrollContainer', type: ['string','Element'], default: 'table wrapper', description: 'Selector or HTML Element to the scroll container. If table wrapper\'s height is not set to 100% of the container, and is taller than the container - the container will have to be scrollable, and in this case it must be provided here.' },
@@ -60,7 +61,7 @@ const apiProps = <ApiProp[]>[
 	{ name: 'interactive', type: 'boolean', description: 'Makes table rows selectable with mouse and adds keyboard navigation.' },
 	{ name: 'multiselect', type: 'boolean', description: 'If true - a column with checkboxes will be added to the grid.' },
 	{ name: 'sortby', type: 'string', description: 'Field name to sort by on initial render.' },
-	{ name: 'sortdir', type: 'string', default: '<em>ASC</em>', description: 'Initial sort direction. One of <em>ASC</em> or <em>DESC</em>.' },
+	{ name: 'sortdir', type: ['ASC', 'DESC'], default: '<em>ASC</em>', description: 'Initial sort direction.' },
 
 	{ name: 'onclick', type: 'function', description: 'Triggered after a row has been clicked.' },
 	{ name: 'ondblclick', type: 'function', description: 'Triggered after a row has been double-clicked.' },
@@ -68,14 +69,24 @@ const apiProps = <ApiProp[]>[
 	{ name: 'onselect', type: 'function', description: 'Triggered after a row selection has changed.' },
 ];
 
+const columnProps = <ApiProp[]>[
+	{ name: 'field', type: 'string', required: true, description: 'The field name of the column (must correspond to a key in the data objects).' },
+	{ name: 'label', type: 'string', description: 'The label of the column.' },
+	{ name: 'width', type: 'number', description: 'The width of the column.' },
+	{ name: 'align', type: ['left', 'center', 'right'], description: 'The alignment of the column.' },
+	{ name: 'sortable', type: 'boolean', description: 'Whether the column is sortable.' },
+	{ name: 'total', type: 'boolean', description: 'Whether the column should display a total.' },
+	{ name: 'renderer', type: 'function', description: 'A function to render the column content.' },
+];
 
 
-const columns = [
+
+const columns: GridColumn[] = [
 	{ field: 'name', label: 'Name', sortable: true,
 		renderer: (item) => `<a href="#Grid/${item.id}">${item.name}</a>`
 	},
-	{ field: 'date', label: 'Date', width: 200, sortable: true },
-	{ field: 'price', label: 'Price', width: 200, sortable: true, total: true,
+	{ field: 'date', label: 'Date', width: 110, sortable: true },
+	{ field: 'price', label: 'Price', width: 110, align: 'right', sortable: true, total: true,
 		renderer: (item) => `€${item.price}`
 	}
 ];