+++ src/grid/parts/GridFoot.svelte
@@ -4,7 +4,7 @@
 			<td></td>
 		{/if}
 		{#each $columns as column (column.field)}
-			<td class="td-{getType(column)}">
+			<td class="cell-align-{column.align || 'left'}">
 				{@html renderer(column)}
 			</td>
 		{/each}
@@ -12,29 +12,18 @@
 </tfoot>
 
 <script lang="ts">
-import type { DataStoreType } from '../types';
-
-
-interface Props {
-	multiselect?: boolean;
-	Data?: DataStoreType
-}
+import type { GridPartProps } from '../types';
 
 const {
 	multiselect = false,
 	Data
-}: Props = $props();
+}: GridPartProps = $props();
 
 const columns = $derived(Data.columns);
 
 
 function sumColumn (column) {
 	return $Data.reduce((acc, row) => acc + +row[column.field], 0);
-}
-
-
-function getType (column) {
-	return typeof $Data?.[0]?.[column.field];
 }
 
 function renderer (column) {